arregle la estrucutra del proyecto. cre un componente input que ya manda la informacion al sever.
