--- frontend/src/home/page/galeria de compra/.txt
@@ -1,9 +1,0 @@
-aqui me encargo yo 
-
-css
-header
-main 
-footer
-
-components
-
