--- frontend/src/home/page/home page/.txt
@@ -1,8 +1,0 @@
-entro yo y rinaldi 
-
-css
-footer
-header 
-main 
-
-eso es lo que falta
