--- frontend/src/home/page/login/login-clie.css
@@ -1,233 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@100;300&display=swap");
-@import url('https://fonts.googleapis.com/css2?family=Lobster+Two:ital@1&display=swap');
-
-* {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-    text-decoration: none;
-    font-family: Poppins;
-}
-
-body {
-    background-color: #194c99;
-}
-
-.container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 80vh;
-    padding: 40px 60px 120px;
-    margin: auto;
-}
-
-.container-img {
-    display: flex;
-    justify-content: space-evenly;
-    background-image: linear-gradient(
-        rgba(0, 0, 0, 0.7),
-        rgba(0, 0, 0, 0.7)),
-        url('../assets/fondo.webp');
-
-    margin-left: 200px;
-    padding: 40px 60px 80px;
-    height: 600px;
-    width: 660px;
-    border-radius: 10px 0 0 10px;
-}
-
-.container-img .title {
-    color: #FFF ;
-}
-
-.container-img h1 {
-    font-size: 90px;
-    text-align: left;
-    font-weight: 700;
-}
-
-.container-img .social-icon {
-    position: relative;
-}
-
-.input-groupt {
-    background-color: rgb(255, 255, 255);
-    padding: 20px 40px 60px;
-    height: 600px;
-    width: 495px;
-    border-radius: 0 10px 10px 0;
-}
-
-.register-link {
-    margin: 40px;
-}
-
-.register-link a {
-    color: #3d5dee;
-}
-
-.fol {
-    color: #3d5dee;
-}
-.input-groupt h1 {
-    margin: 10px;
-    padding: 12px;
-    justify-content: space-evenly;
-    font-size: 60px;
-    font-weight: 900;
-}
-
-.input-feild {
-    margin: 0 auto;
-    height: 80px;
-}
-
-.input-feild input {
-    background-color: aliceblue;
-    padding: 8px;
-    width: 100%;
-    color: rgb(175, 175, 175);
-    border: 1px solid rgba(0, 0, 0, 0.438);
-    border-radius: 10px;
-    outline: none;
-}
-
-.input-feild label {
-    position: relative;
-    display: inline-block;
-    top: -45px;
-    left: 0px;
-    color: black;
-    margin: 12px;
-    transition: all cubic-bezier(0.215, 0.610, 0.355, 1) 0.5s;
-}
-
-input:focus~label {
-    transform: translateY(-20px);
-    font-size: 13px;
-    background-color: rgb(255, 255, 255);
-    border: none;
-    border-radius: 4px;
-    padding: 5px;
-    width: 50px;
-}
-
-input:focus~.password {
-    transform: translateY(-21px);
-    font-size: 13px;
-    background-color: rgb(255, 255, 255);
-    border: none;
-    border-radius: 4px;
-    padding: 5px;
-    width: 70px;
-}
-
-input:focus {
-    border: 2px solid rgb(0, 0, 0);
-}
-
-input:checked {
-    background-color: cadetblue;
-
-}
-
-.remenber {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    width: 100%;
-}
-
-.remenber .ramen {
-    padding-left: 12px;
-}
-
-.remenber.remenber-1 input[type="checkbox"]:checked {
-    margin: 12px;
-
-}
-
-.button button {
-    position: absolute;
-    color: #fff;
-    background-color: #194c99;
-    font-size: 16px;
-    font-weight: 500;
-    right: 400px;
-    top: 590px;
-    padding: 10px;
-    margin: 15px;
-    width: 100px;
-    border: none;
-    border-radius: 12px;
-}
-
-button:hover {
-    border-color: #646cff;
-    cursor: pointer;
-}
-
-.container-img h1 {
-    margin: 20px;
-    font-size: 60px;
-    font-weight: 100px;
-    font-family: 'Lobster';
-}
-
-.container-img .texto {
-    margin-top: 40px;
-    margin-left: 20px;
-    font-size: 18px;
-    color: #fff ;
-}
-
-.container-img .social-icon {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    top: 430px;
-    right: 300px;
-}
-
-.social-icon .img {
-    margin-top: 18px;
-    margin-left: 20px;
-    text-align: center;
-    cursor: pointer;
-}
-
-.social-icon .insta {
-    display: flex;
-    font-size: 20px;
-    width: 60px;
-    height: 60px;
-    margin: 10px;
-    background: rgb(156,0,105);
-    background: linear-gradient(142deg, rgba(156,0,105,1) 10%, rgba(255,180,0,1) 39%, rgba(246,9,153,1) 77%);
-    color: white;
-    border-radius: 50px;
-    border: none;
-}
-
-.social-icon .face {
-    display: flex;
-    font-size: 20px;
-    width: 60px;
-    height: 60px;
-    margin: 10px;
-    border-radius: 50px;
-    border: none;
-    color: #fff;
-    background-color: #194c99;
-}
-
-.social-icon .twi {
-    display: flex;
-    width: 60px;
-    height: 60px;
-    margin: 10px;
-    border-radius: 50px;
-    border: none;
-    font-size: 20px;
-    color: #fff;
-    background-color: #00b5ec;
-}
