--- frontend/src/home/page/reserva/.txt
@@ -1,6 +1,0 @@
-esto es de dary 
-
-css
-components
-
-y todo lo necesario para que crees la reserva
