+++ frontend/src/system/dashboard/css/table.css
@@ -1,0 +1,89 @@
+/*? table */
+
+.content {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 95vh;
+  }
+  
+  .content .content-table {
+    padding: 10px 40px 40px;
+    background-color: rgba(149, 194, 231, 0.301);
+    backdrop-filter: blur(10px);
+    box-shadow: 0 0.4rem 0.8rem #0005;
+    border-radius: 0.8rem;
+  }
+  
+  table {
+    background-color: rgba(149, 194, 231, 0.233);
+    border-collapse: collapse;
+    border-spacing: 12px;
+  }
+  
+  tr th {
+    text-align: left;
+  }
+  
+  td img {
+    width: 40px;
+    height: 45px;
+    margin-right: 1rem;
+    border-radius: 50%;
+    vertical-align: middle;
+  }
+  
+  table,
+  th,
+  td {
+    padding: 15px;
+  }
+  
+  table tr:nth-child(even) {
+    background-color: #24da8525;
+  }
+  
+  .navigate-table button {
+    background-color: rgb(56, 207, 81);
+    color: white;
+    font-size: 15px;
+    font-weight: 400;
+    padding: 5px;
+    width: 120px;
+    height: 40px;
+    margin: 15px;
+    border: none;
+    border-radius: 12px;
+    transition: background-color 0.25s cubic-bezier(0.47, 0, 0.745, 0.715);
+  }
+  
+  .navigate-table button:hover {
+    transform: scale(1.025);
+    background-color: rgba(77, 77, 228, 0.651);
+  }
+  
+  
+  .navigate-table label{
+      background-color: white;
+      border: none;
+      border-radius: 5px;
+      position: relative;
+      font-size: 15px;
+      top: 5px;
+      left: 20px;
+      padding: 5px;
+  
+  }
+  .navigate-table input {
+    border: none;
+    border-radius: 5px;
+    padding: 9px;
+    background-color: white;
+    outline: none;
+  }
+  
+  tr:hover td {
+    background-color: rgba(255, 255, 255, 0.363);
+    border-color: #000;
+  }
+
